Add snacraft config
This allows the easy generation of snaps as requested in #227.

This commit does not add automatic generation and publication of snaps
to the Snap Store, but they can easily be generated for local/personal
use by

```
snapcraft
sudo snap install gutenberg_xxx.snap
```

Add directories with snapcraft artifacts to `.gitignore`.