Attempt to fix snap package clean build
The original snap recipe lacks a few build dependencies and thus errors in cleanbuild, this patch fixes it by enumerated them in `build-packages`.

Also a seemingly bug in the rust plugin is encountered and is workarounded by using a third party plugin implementation by @tismith.

NOTE: At my end the build still fails frequently due to retry count depletion when downloading crates.io libraries, which is not resolved in this patch.

Refer-to: Problem building rust based snap - snapcraft - snapcraft.io <https://forum.snapcraft.io/t/problem-building-rust-based-snap/5776>
Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>