Edit docs to clarify RSS process
Add details to the templates/RSS page to clarify that the `rss.xml`
page is only generated if the `generate_rss = true` variable is set
in the site's `config.toml` page.  This information is already present
in other parts of the documentation, but is not present in the RSS page,
which could be confusing.