Edit docs to clarify role of static folder
The existing docs briefly mentions the static folder under `directory
structure`, by stating that the static folder can "contain[] any kind of
files. All the files/folders in the static folder will be copied as-is in
the output directory."

However, the role of the static folder is not mentioned when discussing
how to add content to the site.  This commit adds a short section to the
`Content -- overview` section explaining how to use the static folder to
add static content and providing a few use cases.