Add `index page` section to documentation
The current documentation does not describe how to create a index page.
I initially found this confusing, because I expected an index page to be
a **page** rather than a section.  Thus, I tried to access the page
content with `{{ page.content }}` and was very frustrated when I could
not.

This addition clarifies that the index page is **always** a section,
even if it does not have any sub-pages.  This should also help people
who intend to use Gutenberg to build stand-alone webpages, rather than
blogs.