Add a link to a live demo for the After Dark theme
After hearing no objection to the plan I suggested in
keats/gutenberg #332, I went ahead and put together a live site that
demos the After Dark theme (I plan to do them all, and started with
After Dark simply because it comes first alphabetically).  The demo site
is available at <https://after-dark--gutenberg-theme-demo.netlify.com/>.

One reason I think that live demos would be helpful is that Hugo tends
to have them (see, e.g, <https://themes.gohugo.io/theme/after-dark/>),
and I found them to be much more helpful in getting a feel for a theme
than a screenshot or description is.

Happy to make any changes to this commit or to the live demo if you
would like.