Edit docs to say that sections require _index.md
The prior version of the documentation stated that sections are created
automatically, but this was not correct: sections are created only if
the folder contains an `_index.md` file.  This commit changes the
documentation to reflect that reality and updates the surrounding
language for consistency and to explain the behaviour of `index.md`
files.