Remove www.bharatkalluri.in from EXAMPLES.md
The footer and source code at <https://www.bharatkalluri.in/> state that
the site was made with Hugo, so it should be deleted as an example of a
Gutenberg site.

I considered also removing the link to https://adrien.is/, since the
link appears to be dead.  But I left it for now, since that could be a
temporary outage.

If you would like, I'd be happy to format EXAMPLES.md as a table, which
might help keep it looking clean as it gets more entries.