Adds external highlighters for code blocks.
This feature is activated by specifying shell commands in config.toml.
For example, an external "highlighter" that renders GraphViz to SVG
could be defined as:

    # ...
    [external_highlighters]
    dot = "echo -n '<p>'; dot -Tsvg; echo '</p>'"
    # ...

A code block like:

    ```dot
    digraph G {
	A -> B;
	B -> C;
	A -> C;
    }
    ```

then results in the HTML output by piping the contents of the code block
to the command specified as external_highlighters.dot.

The following environment variables are passed to external highlighter
commands:

- ZOLA_BASE_URL: The base URL in the config
- ZOLA_CURRENT_PAGE: The permalink URL to the current page
- ZOLA_OUTPUT_DIR: The directory being output to

The last one would be more useful if the output directory weren't
cleared before the content is translated from Markdown (ref
https://zola.discourse.group/t/230). This would be particularly useful
for e.g. generating images from LaTeX, and replacing the code blocks
with <img> tags.