Further refine the slugification story
https://github.com/getzola/zola/issues/768#issuecomment-570557713 is the
discussion and source of what I’ve done in this commit.

The work included in here:

- Stop stripping any characters from paths or anchors: none of the
  restrictions are fundamentally valid, and they prevent realistic uses.

- Split slugify_paths into slugify_paths and slugify_anchors, and change
  anchors to using slugify_anchors.

- Turn slugify_paths off by default, and ensure that all the
  documentation is consistent about this (it wasn’t).

One potentially controversial part of this is ceasing turning spaces
into underscores in anchors with slugify_anchors off. I consider that a
legitimate slugification scheme, and contrary to the declared intent of
`slugify_anchors = false`. I think it’s best left to the future where we
can specify `slugify_anchors = "Spaces_to_underscores"` or similar.
(Because there are further variants of it.)