Split slugification of taxonomy terms out
I realised that there can be a fairly big semantic difference between
page/section paths and taxonomy terms, for similar reasons to why
anchors are different.

I believe that defaulting slugify_paths to false is the right thing to
do, but not if it covers taxonomy terms as well—you *do* want them to be
slugified by default.

This means that the breaking change from 0.9 is now very minimal.