Add feature flags to reduce dependencies
We have 223 unique external direct and transitive dependencies.  Many
of these are needed only for serving files via HTTP.  Others are
needed only for making outbound HTTP/S requests.  In certain
situations, such as when building Zola as part of a continuous
integration pipeline, these features are unnecessary, and we would
prefer a faster and simpler build.

In this commit, we add `serve` and `request` feature flags to Zola.
The `serve` flag controls whether the embedded HTTP server is built
and whether the `zola serve` command is available.  The `request` flag
controls whether we support making outbound HTTP/S requests, such as
for link checking.  Both of these features are enabled by default.

With both of these features disabled, such as when building with
`--no-default-features`, we require only 160 unique external
dependencies, a 28% reduction.  Notably, we're able to drop both
`openssl` and `tokio`.  Dropping `openssl` in particular can help with
producing statically-linked builds.

With the `serve` flag, we depend on 210 unique external crates.  With
the `request` flag, we depend on 199 unique external crates.