Use assets in `static/` as page assets
Before this change, assets collocated with a page in `content/` are
made available in `page.assets` and `page.serialized_assets`, but not
assets in `static`.

This change adds assets in `static`, searched in the same way as
collocated assets, to `page.assets` and `page.serialized_assets`.

For example, given a page in `content/a/b/index.md`, all assets in
both `content/a/b` **and** `static/a/b` will be included in
`page.assets` and `page.serialized_assets`.

Signed-off-by: Simão Mata <sm@0io.eu>