Fix RSS block example in documentation (#1210)
This commit fixes a bug in the documentation for the feeds page which
shows how to use the RSS block to enable feed autodiscovery.

The bug used an em space character instead of a space in one part of the
code snippet. If a user were to copy the code snippet into their code as
directed, the Zola build would fail with a parse error. The em space
appears identical to a regular space in monospaced fonts, making the
error seem mysterious or incorrect.

I believe the em space was used in order to prevent the shortcode from
rendering, as the code snippet is just meant to show what the shortcode
looks like. However, it is possible to escape the shortcode so that it
renders correctly without causing confusion for the user who expects to
be able to copy and paste it.

This commit replaces the em space in both code snippets with regular
spaces and escapes the shortcodes.