Provide the current mode to templates
This is useful specifically if you have stuff in your templates such as
`__tera_context` dumping gated behind a configuration setting. It was
previously impossible to tell if the site was building in production to
stop this from ever accidentally getting into a production build even if
the config setting were accidentally enabled.

Providing this to templates solves this issue, although it is impure.