Support standard byte order sorting
This PR adds the `sort_order` property to sections that defaults to
lexical in order to keep backwards compatability and adds "standard" for
byte order sort.

This is useful if you have a list of non-english names and you want them
sorted correctly. My site is in Swedish and with natural sort åäö (our
last three letters of the alphabet) end up in the same position as aao
respectively.