docs(readme) add 1 step install build serve +links
A community of devs in my university and I tried to follow the guide, and we decided, that as much as we love this project, to improve the install, build and serve guide in one command to let newcomers to try and optimize experience as fast as they can, to adopt in mass, faster. Also we added a table with links down bellow to allow devs to know where and how to install zola as fast as they can to plug and play!