inject site config during Sass compilation
it is helpful to be able to use site config during Sass compilation,
for example to allow a site owner to set header color in a Sass
template.  in this change we serialize a subset of the site's config
to a `.scss` file as a Sass map literal, allowing config content to
be referenced in site Sass or theme Sass.

this commit involves bumping the `grass` dependency to version
`1.13.0`.  we do this so that we can use nested map retrieval in
Sass files, such as `map.get($config, extra, background_color)`.