Add `insert_anchor_links` back to `config.toml`
`insert_anchor_links` was a valid `config.toml` setting before [0.0.6]
but was changed to a section option.

[0.0.6]: https://github.com/getzola/zola/blob/master/CHANGELOG.md#006-2017-05-24

It was mentioned in the documentation for `config.toml` until 2021
(#1477).

In 2021, @Keats said that "`insert_anchor_links` could probably be a
config option in `config.toml`":
https://github.com/getzola/zola/issues/634#issuecomment-882775151

In the same thread, several other users requested this feature.
@Artoria2e5 said:

> These sort of settings become annoying to copy-paste when a site ends
> up with many sections.

@ctron said:

> But even more annoying is that it seems I can't add
> insert_anchor_links on the top section, but need to place it in every
> section.
>
> I think having this as a default configuration for the site makes
> total sense.