Fix footnotes semantic organization for accessibility (#2890)
Footnotes are currently excluded in certain accessibility tools such as
Firefox Reader View. This is because footnotes are nested underneath an
HTML5 `footer` element, which permits the accessibility tool to discard
the contents as irrelevant to the current section text under
consideration.

Semantically, the `footer`[^1] element represents metainformation for
its nearest ancestor sectioning content (e.g. publication timestamps,
related links, authorship copyright, etc.). It's incorrect to place
footnotes underneath the `footer` element because, while footnotes may
be _thematically_ distinct from the main content, they are considered
**semantically** part of that content. Footnotes are merely a different
_section_ of content: their role is that of content supplementation, not
metainformation.

In order to properly indicate the semantic relationship between the
main content and its footnotes, the footnotes should be nested
underneath a `section`[^2] element which adequately communicates this
relationship. This isn't just speculation, but is in fact the exact way
the WHATWG HTML5 specification document recommends implementing the
footnotes with reciprocal backlinks idiom.[^3]

To that end, this patch changes the footnotes `footer` element to a
`section` element. The `class="footnotes"` attribute remains the same so
that user stylesheets may not be affected by this bug fix:

    <section class="footnotes">
    <ol class="footnotes-list">
    ...
    </ol>
    </section>

Test snapshots are also updated to reflect the adjustment.

[^1]: https://html.spec.whatwg.org/multipage/sections.html#the-footer-element
[^2]: https://html.spec.whatwg.org/multipage/sections.html#the-section-element
[^3]: https://html.spec.whatwg.org/multipage/semantics-other.html#footnotes

Fixes: c4a84a8e ("Improve accessibility by nesting bottom footnotes inside footer element (#2688)")
Link: https://github.com/getzola/zola/pull/2688

Co-authored-by: William Breathitt Gray <wbg@losingoneself.com>