Explicitly prohibit --debounce 0
The `tick_rate` at which the debouncer library checks whether any events
should now be dispatched shouldn't be 0.  It cannot either be any
greater than the debounce rate itself.

Ideally, we may want to *disable* debouncing altogether when the user
passes --debounce 0, but that'd require a special case to be made in the
instrumentation, where we'd bypass the debouncer: that's another package
to depend on, and ultimately deemed no worth the hassle.

Ref: https://github.com/getzola/zola/issues/2937
Ref: https://zola.discourse.group/t/watch-debounce-configuration/2666