Configurable file watcher debounce duration in live-reload mode (#2938)
* Introduce -d|--debounce <VALUE> flag for serve

Let the file watcher use a configurable debounce value.

This commit preserves the (until now hard-coded) 1-second debounce value
as the default, but adds a new flag for the `serve` command, which lets
the user configure a debounce duration for consecutive rebuilds in watch
mode.

Closes #2937.

Ref: https://github.com/getzola/zola/issues/2937
Ref: https://zola.discourse.group/t/watch-debounce-configuration/2666

* Document serve's --debounce flag in cli-usage.md

Ref: 6737349447f79314c15fdd860b26eeb899f38d86
Ref: https://github.com/getzola/zola/issues/2937
Ref: https://zola.discourse.group/t/watch-debounce-configuration/2666

* Explicitly prohibit --debounce 0

The `tick_rate` at which the debouncer library checks whether any events
should now be dispatched shouldn't be 0.  It cannot either be any
greater than the debounce rate itself.

Ideally, we may want to *disable* debouncing altogether when the user
passes --debounce 0, but that'd require a special case to be made in the
instrumentation, where we'd bypass the debouncer: that's another package
to depend on, and ultimately deemed no worth the hassle.

Ref: https://github.com/getzola/zola/issues/2937
Ref: https://zola.discourse.group/t/watch-debounce-configuration/2666

* Fix formatting style

The line was too long.  I pruned the /*tick_rate=*/ inline comment that
was indicating the name of the argument it was annotating, since it was
*the exception*: there are zero other occurrences of such notation
throughout this codebase.