Try to preserve ICC color profiles when processing images
For all the supported input and output formats,
this copies the ICC data (color profile) of the
input image to the output image in `resize_image`.
Discarding the ICC data is generally *wrong*, as
it reinterprets the image data in a different
color space (usually sRGB/REC.709).

Caveats:
- Some formats *may* store color profiles in a
  different way. ICC is the industry standard,
  supported in all of our output formats, and as
  far as I can tell it’s even used for all the
  common color profiles like sRGB and AdobeRGB
- AVIF writer doesn’t support ICC. We print a
  warning in this case.
- Only the `image`-builtin WebP writer supports
  ICC, but it can only compress losslessly.
  Therefore, we only embed ICC if the compression
  is lossless (using a different writer from what
  we normally do). We print a warning otherwise.

Also adds tests with a few common color profiles
and some of the existing images (which all don’t
have ICC data).